FIX : Various FIX : Failed (so far) attempt to use non EPSG:4326 projections with WMS
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@14497 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ src/Preferences/WMSPreferencesDialog.cpp
@@ -34,7 +34,7 @@ void WMSPreferencesDialog::addServer(const WmsServer & srv)
 	theWmsServers.push_back(srv);
 	if (!srv.deleted) {
 		QListWidgetItem* item = new QListWidgetItem(srv.WmsName);
-		item->setData(Qt::UserRole, theWmsServers.size()-1);
+		item->setData(Qt::UserRole, (int)(theWmsServers.size()-1));
 		lvWmsServers->addItem(item);
 	}
 }
